In debug mode we're getting a segfault when dispatch the Display
I believe the reason for this is that the release-small build is
managing to only allocate a single Display on the stack.

However in debug it'll be allocating one on the stack in init
then copying it into the stack frame above. Therefore the epoll
event will point to the init dispatchable which will become invalid.

I wonder if there's another way around this, but separate out
adding to epoll to a function addToEpoll that will take a
pointer to our actual Display