WIP commit: save some intermediate work for the historical process
The implementation of draw is taking a pointer to Self and seems to
work. The switch is then on (self.*) and each branch of the switch
binds a |*<var>|.

But I think it still works if I remove all the pointer stuff which
I don't quite get.

If it takes just Self I would expect it to not compile