We might have a situation where a client has hungup whilst we're already processing messages from it. If that is the case we won't have HUP. But when we attempt to send a message we will get error.BrokenPipe.
That's fine, we can use that as the hangup signal, but
we need some way to stop processing messages because the
client will already have been garbage collected.

We can use the fact that even though we have collected the
client, we will not have gotten round to allowing another
connecting client to use its pool slot and so can set a
tombstone flag on the client, which we can use to then
stop processing messages.