Latest zig + epoll iterator (#67)
# Description

- Elm-architecture-inspired event iterator
- The previous version populated the epoll event data with a pointer.
This pointer was a function pointer to an implementation of a wayland
message handler. When processing messages / events this then happens in
a relatively hidden fashion. Particularly code navigation will be
hampered and the code generally harder to understand.
- This change replaces the function pointer dispatch with a data /
iterator approach. When we have activity on some subsystem(s) epoll will
wake up and return an event iterator. Each call to `.next()` will return
a new event (or null). We will exhaust all events for a given subsystem
before moving onto another subsystem that has activity.
- We exchange a function call (with the function pointers) with
returning a data structure for each event.
    - Those events are then dispatched by `switch`ing on the event.
- In this way we end up with a very straightforward dispatch method that
is very explicit in the code and easy to reason about / navigate over.
- Resource allocation has been reworked.
- Previously resources (clients, window, regions, etc.) were statically
allocated.
- This has been replaced with `IterablePool` and `SubsetPool`. These
allocate on startup but not thereafter.
- The wayland protocol implementation is completely rewritten to provide
way better type safety with an explicit `union` of wayland objects and
messages.
- The protocol generator has been rewritten in zig (using
https://github.com/mitchellh/zig-build-libxml2 for XML parsing) so that
we can jettison the `python` dependency. The generator is currently at
https://github.com/malcolmstill/foxwhale-gen. I think I'll move the
generator code into this repo, but will wait for
https://github.com/mitchellh/zig-build-libxml2/pull/3 so I don't have to
vendor to include the changes in that PR

# Closes
- Closes https://github.com/malcolmstill/foxwhale/issues/59
- Closes https://github.com/malcolmstill/foxwhale/issues/58
- Closes https://github.com/malcolmstill/foxwhale/issues/44
- Closes https://github.com/malcolmstill/foxwhale/issues/11

# Follow ups

- Reimplement DRM/KMS backend